+++ buildSrc/src/test/groovy/com/google/firebase/gradle/plugins/measurement/GenerateMeasurementsTaskTest.groovy
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.gradle
+package com.google.firebase.gradle.plugins.measurement
 
 import static org.junit.Assert.assertFalse
 import static org.junit.Assert.assertTrue
